requirements.txt
matplotlib==3.8.2
moviepy==1.0.3
opencv_contrib_python==4.8.0.74
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
pandas==1.5.2
Pillow==10.1.0
pymongo==4.5.0
python-dotenv==1.0.1
streamlit==1.28.0
twilio==9.0.0
ultralytics==8.1.24